Refactor null checking in Observable methods
This commit refactors the null checking in the methods of the Observable class to use nullish coalescing operator. The changes also affect some tests, where the condition is explicitly set to undefined rather than any type 0. This makes the code more succinct and more accurately represents cases where the condition is not provided.