Update Filter class and version in package.json
A new response object has been added to the Filter class in Observables. This modification includes redefining the processChain method to make use of the new response object. It ensures the isOK attribute and payload are reset to their default states respectively. The application version has been incremented to 1.9.49 in package.json file.