Update callback functions and increase version
This commit modifies the structure of several callback functions in Observable, Pipe, and FilterCollection to utilize the new IFilterChainCallback and IChainCallback data types. This improved the separation of concerns and code readability. Additionally, an extra test for unsubscribing has been added to ensure functionality. The package version has also been stepped up.